Adding pemain module

+++ app/Models/Tim.php
@@ -37,14 +37,31 @@ class Tim extends Model
         'soft_delete',
     ];
 
-    public function go_list()
-    {
-        $data = Tim::paginate(10);
+    public function go_list() {
+        
+        $data = Tim::select(
+            'tim.id_perusahaan',
+            'tim.nama',
+            'tim.logo',
+            'tim.tahun_berdiri',
+            'tim.alamat_markas_tim',
+            'tim.kota_markas_tim',
+            'tim.created_at',
+            'perusahaan.nama',
+        )
+        ->where('soft_delete')
+        ->join('perusahaan', 'perusahaan.id', 'tim.id_perusahaan')
+        ->paginate(10);
+
         return $data;
     }
 
-    public function go_create($input)
-    {
+    public function go_exists($id) {
+        $data = Tim::where('id', $id)->exists();
+        return $data;
+    }
+
+    public function go_create($input) {
         $data = Tim::create([
             'id_perusahaan' => $input->id_perusahaan,
             'nama' => $input->nama,
@@ -56,17 +73,40 @@ class Tim extends Model
         return $data;
     }
 
-    public function go_update($input, $id)
-    {
+    public function go_update($input, $id) {
         $data = Tim::where('id', $id)->update([
-            'nama' => $input->nama
+            'id_perusahaan' => $input->id_perusahaan,
+            'nama' => $input->nama,
+            'tahun_berdiri' => $input->tahun_berdiri,
+            'alamat_markas_tim' => $input->alamat_markas_tim,
+            'kota_markas_tim' => $input->kota_markas_tim,
         ]);
         return $data;
     }
 
-    public function go_delete($id)
-    {
-        $data = Tim::where('id', $id)->delete();
+    public function go_delete($id) {
+        $soft_delete = Tim::select('soft_delete')->where('id', $id)->first()->soft_delete;
+        
+        if ($soft_delete) {
+            $data = Tim::where('id', $id)->delete();
+            return $data;
+        }
+
+        $data = Tim::where('id', $id)->update([
+            'soft_delete' => true
+        ]);
+        return $data;
+    }
+
+    public function go_update_logo($id, $file) {
+        $data = Tim::select('logo')->where('id', $id)->update([
+            'logo' => $file
+        ]);
+        return $data;
+    }
+
+    public function go_get_old_logo($id) {
+        $data = Tim::select('logo')->where('id', $id)->first()->logo;
         return $data;
     }
 }
